Serialize empty QKeySequence object as none
When doing e.g. KGlobalAccel::setGlobalShortcut(QKeySequence()) we
get QList(QKeySequence()) for defaultShortcuts.

This serializes as an empty string. Functionally it's the same as
no sequence though, so serialize it as 'none'.

Otherwise we are in a situation where we sometimes write "" and sometimes
"none" to the file, which is weird